+++ config_start_daily_run.bat
@@ -1,2 +1,2 @@
-@CALL "C:\ProgramData\Anaconda3\Scripts\activate.bat" Anaconda3
-@CALL python C:\Users\haoli\Desktop\python\Investment\HFinvest\launcher.py
+@CALL "C:\ProgramData\Anaconda3\Scripts\activate.bat" C:\ProgramData\Anaconda3
+@CALL python G:\Projects\HFinvest\top_daily_run.py
